# examples/hono-user-admin/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hono user/admin authorization example",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://127.0.0.1:3000" }],
  "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
  "paths": {
    "/users/{userId}": {
      "get": {
        "operationId": "getUser",
        "summary": "Read a user resource",
        "description": "Users may read themselves. Admins may read any user. Requests by a different normal user return 404 so the resource's existence is not disclosed.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateUser",
        "summary": "Update a user resource",
        "description": "Users may update themselves. Admins may update any user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["displayName"],
                "properties": { "displayName": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user",
            "headers": {
              "X-Resource-State": {
                "schema": { "type": "string", "const": "updated" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": {
            "description": "The display name is invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/audit": {
      "get": {
        "operationId": "getAdminAudit",
        "summary": "Read the admin-only audit summary",
        "tags": ["admin"],
        "responses": {
          "200": {
            "description": "The audit summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["activeUsers"],
                  "properties": {
                    "activeUsers": { "type": "integer", "minimum": 0 }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Synthetic bearer tokens returned by the test fixture endpoint"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Synthetic server-side user session"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "displayName": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid bearer token or session cookie",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated caller lacks the admin role",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "The user does not exist or is hidden from this caller",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}